backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
pydantic>=2.6.1
python-dotenv>=1.0.0
requests>=2.31.0
deepgram-sdk>=2.12.0
numpy>=1.26.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-magic>=0.4.27
aiofiles>=23.2.1
yt-dlp>=2024.3.10
playwright>=1.42.0
selenium>=4.18.1
pytest>=8.0.0 